requirements.txt
uvicorn
scikit-learn==1.8.0
joblib
pydantic
nltk
scipy
numpy
streamlit
requests
python-dotenv
fastapi
uvicorn
python-dotenv
lightgbm
